feat(grpcgen/message): handle underlying types of `*types.Named`
A `*types.Named` can be one of:

- *types.Struct
- *types.Interface
- *types.Basic

In the case of a struct, we'll need to parse the fields into a `Message`.

In the case of an interface, we'll have to find all of its implementations.

In the case of basic types, we need to check if they are enums, and if so,
find all of their possible values.

Note that the only time `MessageTypeConverter.Convert` can receive an instance
of `*types.Interface` is if the type is `interface{}`. Otherwise, the type
would have been a `*types.Named` with an underlying `*types.Interface`.
Therefore it should be impossible to get a non-empty interface.