feat(codegen): generate messages
This patch implements the bulk of the protobuf generation, with the
exception of enums and interfaces. There might be some builtin Go types
that we'll need to replace to have a better representation. Currently,
`error` is replaced with an optional string. We might have to do similar
stuff for `Time`, etc.